dates: force UTC timezone
In the RFC, it is stated that all (or almost all) dates should be
"UTCDate" (e.g. in the UTC timezone and formatted with a trailing "Z").

Convert the dates to use the UTC timezone when required before
marshaling to JSON. Create type aliases to be able to call
json.Marshal() from MarshalJSON() without a recursion error.

Link: https://www.rfc-editor.org/rfc/rfc8620.html#section-1.4
Link: https://www.rfc-editor.org/rfc/rfc8621.html#section-4.4.1
Link: https://choly.ca/post/go-json-marshalling/
Signed-off-by: Robin Jarry <robin@jarry.cc>